Minor refactor :wink:
